+++ course-app/src/components/new-password/newPassword.js
@@ -1,6 +1,5 @@
-import LogoSvg from "../../svg image/logo-svg/logoSvg";
-import "../loginpage/login.css"
-
+import LogoSvg from "../../svg_image/logo-svg/logoSvg";
+import '../../pages/sign_inup/login.css'
 
 
 export default function NewPassword () {
