+++ course-app/src/pages/course_info/course.css
