+++ course-app/src/pages/main/main.css
@@ -145,7 +145,7 @@ body {
     height: 48px;
     border-radius: 46px;
     background-color: #c7e957;
-    margin: 0 20px auto auto;
+    margin: 0 auto 20px;
     display: block;
 
     color: #000;
