+++ course-app/src/pages/main/mainPage.js
