+++ course-app/src/svg_image/arrow-svg/arrowSvg.js
