+++ course-app/src/svg_image/logo-svg/logoSvg.js
